refactor: 将路由配置迁移到 routing 模块并统一内容集合
- 删除 docs 项目中的 releases 页面和内容配置
- 将 releases 集合配置迁移到 layer/content.config.ts
- 新增 routing 模块统一管理路由逻辑
- 优化内容源配置,支持动态检测 releases 文件格式
